Select reaction (#3)
Add a selection menu to that displays selected reaction

* Add mostly working menu

* Engine class addition and some code cleanup

Added the engine class that takes the information provided by the data classes and turns it into SceneKit nodes that can then be rendered in ViewController. Currently does not animate or distinguish between bond types. That functionality will be added once bond types are added into the data classes.

* Removed junk

* Minor fix

* Support reading partial bonds

* Update gitignore

* Merge from Engine-Testing

* Add reactions, make reaction accessible

* Add selection menu and more reaction files

* Adds temp(?) fix to reset scene

* Fix missing atoms

* Update scene lighting

* Fixed missing buttons

* Fix blank space

Co-authored-by: jerbrowsky <83240817+jerbrowsky@users.noreply.github.com>